+++ src/fastmath/complex.clj
@@ -22,7 +22,6 @@
             [fastmath.vector :as v])
   (:import [fastmath.vector Vec2]))
 
-(set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 (m/use-primitive-operators #{'abs})