+++ src/fastmath/curves.clj
@@ -5,7 +5,6 @@
             [fastmath.random :as r])
   (:import [fastmath.vector Vec2]))
 
-(set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 (m/use-primitive-operators)