+++ src/fastmath/easings.clj
@@ -15,7 +15,6 @@
                         :cr "Creators"}}
   (:require [fastmath.core :as m]))
 
-(set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 (m/use-primitive-operators)