+++ src/fastmath/fields.clj
@@ -49,7 +49,6 @@
             [fastmath.vector :as v])
   (:import [fastmath.vector Vec2]))
 
-(set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 (m/use-primitive-operators)