+++ src/fastmath/optimization/bo.clj
@@ -5,7 +5,6 @@
             [fastmath.core :as m]
             [fastmath.optimization :as opt]))
 
-(set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 (m/use-primitive-operators)