+++ src/fastmath/stats.clj
@@ -61,7 +61,6 @@
            [org.apache.commons.math3.stat.correlation KendallsCorrelation SpearmansCorrelation PearsonsCorrelation]
            [org.apache.commons.math3.stat.regression SimpleRegression]))
 
-(set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
 (m/use-primitive-operators)